AlgoRewards/projects/AlgoRewards-contracts/pyproject.toml
[tool.poetry]
name = "AlgoRewards-contracts"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["elnicomaldonado <62966960+elnicomaldonado@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^4.0.0"
python-dotenv = "^1.0.0"
algorand-python = "^2.0.0"
algorand-python-testing = "~0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
algokit-client-generator = "^2.1.0"
puyapy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

